Cybersecurity-Agent/requirements.txt
rich
python-whois
dnspython
requests
# ── LLM & Agent framework ────────────────────────────────────────────────────
openai>=1.30.0
langchain-openai>=0.1.8
langchain-core>=0.2.0
langgraph>=0.2.0
langchain-mcp-adapters>=0.1.0

# ── MCP ──────────────────────────────────────────────────────────────────────
mcp>=1.0.0

# ── API server ────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── HTTP client (supervisor → agent) ────────────────────────────────────────
httpx>=0.27.0

# ── GitHub API ───────────────────────────────────────────────────────────────
requests>=2.31.0

# ── Config & utilities ────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.7.0
redis>=5.0.0
streamlit>=1.38.0
langgraph-checkpoint-redis>=0.3.0
